Refactor test folder structure
