+++ test/management/client.tests.js
@@ -1,10 +1,12 @@
 var expect = require('chai').expect;
 var nock = require('nock');
 var Promise = require('bluebird');
-var ClientsManager = require('../src/ClientsManager');
 
-var ArgumentError = require('../src/exceptions').ArgumentError;
+var SRC_DIR = '../../src';
 var API_URL = 'https://tenant.auth0.com';
+
+var ClientsManager = require(SRC_DIR + '/management/ClientsManager');
+var ArgumentError = require(SRC_DIR + '/exceptions').ArgumentError;
 
 
 describe('ClientsManager', function () {
