+++ test/management/connections.tests.js
@@ -1,10 +1,11 @@
 var expect = require('chai').expect;
 var nock = require('nock');
 
-var ConnectionsManager = require('../src/ConnectionsManager');
-var ArgumentError = require('../src/exceptions').ArgumentError;
-
+var SRC_DIR = '../../src';
 var API_URL = 'https://tenant.auth0.com';
+
+var ConnectionsManager = require(SRC_DIR + '/management/ConnectionsManager');
+var ArgumentError = require(SRC_DIR + '/exceptions').ArgumentError;
 
 
 describe('ConnectionsManager', function () {
