+++ test/management/jobs.tests.js
@@ -1,16 +1,20 @@
+var path = require('path');
 var expect = require('chai').expect;
 var nock = require('nock');
+var extractParts = require('../utils').extractParts;
 
-var EmailProviderManager = require('../src/EmailProviderManager');
-var ArgumentError = require('../src/exceptions').ArgumentError;
-
+var SRC_DIR = '../../src';
 var API_URL = 'https://tenant.auth0.com';
 
+var JobsManager = require(SRC_DIR + '/management/JobsManager');
+var ArgumentError = require(SRC_DIR + '/exceptions').ArgumentError;
 
-describe('EmailProviderManager', function () {
+
+describe('JobsManager', function () {
   before(function () {
     this.token = 'TOKEN';
-    this.emailProvider = new EmailProviderManager({
+    this.id = 'testJob';
+    this.jobs = new JobsManager({
       headers: { authorization: 'Bearer ' + this.token },
       baseUrl: API_URL
     });
@@ -18,11 +22,11 @@ describe('EmailProviderManager', function () {
 
 
   describe('instance', function () {
-    var methods = ['configure', 'get', 'update', 'delete'];
+    var methods = ['verifyEmail', 'importUsers', 'get'];
 
     methods.forEach(function (method) {
       it('should have a ' + method + ' method', function () {
-        expect(this.emailProvider[method])
+        expect(this.jobs[method])
           .to.exist
           .to.be.an.instanceOf(Function);
       })
@@ -32,13 +36,13 @@ describe('EmailProviderManager', function () {
 
   describe('#constructor', function () {
     it('should error when no options are provided', function () {
-      expect(EmailProviderManager)
+      expect(JobsManager)
         .to.throw(ArgumentError, 'Must provide client options');
     });
 
 
     it('should throw an error when no base URL is provided', function () {
-      var client = EmailProviderManager.bind(null, {});
+      var client = JobsManager.bind(null, {});
 
       expect(client)
         .to.throw(ArgumentError, 'Must provide a base URL for the API');
@@ -46,7 +50,7 @@ describe('EmailProviderManager', function () {
 
 
     it('should throw an error when the base URL is invalid', function () {
-      var client = EmailProviderManager.bind(null, { baseUrl: '' });
+      var client = JobsManager.bind(null, { baseUrl: '' });
 
       expect(client)
         .to.throw(ArgumentError, 'The provided base URL is invalid');
@@ -55,22 +59,17 @@ describe('EmailProviderManager', function () {
 
 
   describe('#get', function () {
-    var data = {
-      name: 'Test provider',
-      options: {}
-    };
-
     beforeEach(function () {
       this.request = nock(API_URL)
-        .get('/emails/provider')
+        .get('/jobs/' + this.id)
         .reply(200);
     })
 
 
     it('should accept a callback', function (done) {
       this
-        .emailProvider
-        .get(function () {
+        .jobs
+        .get({ id: this.id }, function () {
           done();
         });
     });
@@ -78,8 +77,8 @@ describe('EmailProviderManager', function () {
 
     it('should return a promise if no callback is given', function (done) {
       this
-        .emailProvider
-        .get()
+        .jobs
+        .get({ id: this.id })
         .then(done.bind(null, null))
         .catch(done.bind(null, null));
     });
@@ -89,16 +88,14 @@ describe('EmailProviderManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .get('/emails/provider')
+        .get('/jobs/' + this.id)
         .reply(500);
 
       this
-        .emailProvider
-        .get()
+        .jobs
+        .get({ id: this.id })
         .catch(function (err) {
-          expect(err)
-            .to.exist;
-
+          expect(err).to.exist;
           done();
         });
     });
@@ -107,28 +104,35 @@ describe('EmailProviderManager', function () {
     it('should pass the body of the response to the "then" handler', function (done) {
       nock.cleanAll();
 
+      var data = [{ test: true }];
       var request = nock(API_URL)
-        .get('/emails/provider')
+        .get('/jobs/' + this.id)
         .reply(200, data);
 
       this
-        .emailProvider
-        .get()
-        .then(function (provider) {
-          expect(provider.id)
-            .to.equal(data.id);
+        .jobs
+        .get({ id: this.id })
+        .then(function (blacklistedTokens) {
+          expect(blacklistedTokens)
+            .to.be.an.instanceOf(Array);
+
+          expect(blacklistedTokens.length)
+            .to.equal(data.length);
+
+          expect(blacklistedTokens[0].test)
+            .to.equal(data[0].test);
 
           done();
         });
     });
 
 
-    it('should perform a GET request to /api/v2/emails/provider', function (done) {
+    it('should perform a GET request to /api/v2/jobs', function (done) {
       var request = this.request;
 
       this
-        .emailProvider
-        .get()
+        .jobs
+        .get({ id: this.id })
         .then(function () {
           expect(request.isDone())
             .to.be.true;
@@ -142,17 +146,15 @@ describe('EmailProviderManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .get('/emails/provider')
+        .get('/jobs/' + this.id)
         .matchHeader('Authorization', 'Bearer ' + this.token)
         .reply(200)
 
       this
-        .emailProvider
-        .get()
+        .jobs
+        .get({ id: this.id })
         .then(function () {
-          expect(request.isDone())
-            .to.be.true;
-
+          expect(request.isDone()).to.be.true;
           done();
         });
     });
@@ -161,46 +163,42 @@ describe('EmailProviderManager', function () {
     it('should pass the parameters in the query-string', function (done) {
       nock.cleanAll();
 
-      var params = {
-        include_fields: true,
-        fields: 'test'
-      };
-
       var request = nock(API_URL)
-        .get('/emails/provider')
-        .query(params)
+        .get('/jobs/' + this.id)
+        .query({
+          include_fields: true,
+          fields: 'test'
+        })
         .reply(200)
 
       this
-        .emailProvider
-        .get(params)
+        .jobs
+        .get({ id: this.id, includeFields: true, fields: 'test' })
         .then(function () {
-          expect(request.isDone())
-            .to.be.true;
-
+          expect(request.isDone()).to.be.true;
           done();
         });
     });
   });
 
 
-  describe('#configure', function () {
+  describe('#importUsers', function () {
     var data = {
-      name: 'Test provider',
-      credentials: {}
+      users: path.join(__dirname, '../data/users.json'),
+      connection_id: 'con_test'
     };
 
     beforeEach(function () {
       this.request = nock(API_URL)
-        .post('/emails/provider')
-        .reply(200, data);
+        .post('/jobs/users-imports')
+        .reply(200);
     })
 
 
     it('should accept a callback', function (done) {
       this
-        .emailProvider
-        .configure(data, function () {
+        .jobs
+        .importUsers(data, function () {
           done();
         });
     });
@@ -208,8 +206,8 @@ describe('EmailProviderManager', function () {
 
     it('should return a promise if no callback is given', function (done) {
       this
-        .emailProvider
-        .configure(data)
+        .jobs
+        .importUsers(data)
         .then(done.bind(null, null))
         .catch(done.bind(null, null));
     });
@@ -219,27 +217,25 @@ describe('EmailProviderManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .post('/emails/provider')
+        .post('/jobs/users-imports')
         .reply(500);
 
       this
-        .emailProvider
-        .configure(data)
+        .jobs
+        .importUsers(data)
         .catch(function (err) {
-          expect(err)
-            .to.exist;
-
+          expect(err).to.exist;
           done();
         });
     });
 
 
-    it('should perform a POST request to /api/v2/emails/provider', function (done) {
+    it('should perform a POST request to /api/v2/jobs/users-imports', function (done) {
       var request = this.request;
 
       this
-        .emailProvider
-        .configure(data)
+        .jobs
+        .importUsers(data)
         .then(function () {
           expect(request.isDone())
             .to.be.true;
@@ -249,16 +245,19 @@ describe('EmailProviderManager', function () {
     });
 
 
-    it('should pass the data in the body of the request', function (done) {
+    it('should be a multipart request', function (done) {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .post('/emails/provider', data)
+        .post('/jobs/users-imports')
+        .matchHeader('Content-Type', function (header) {
+          return header.indexOf('multipart/form-data') === 0;
+        })
         .reply(200);
 
       this
-        .emailProvider
-        .configure(data)
+        .jobs
+        .importUsers(data)
         .then(function () {
           expect(request.isDone())
             .to.be.true;
@@ -268,84 +267,127 @@ describe('EmailProviderManager', function () {
     });
 
 
-    it('should include the token in the Authorization header', function (done) {
+    it('should have two parts: connection_id and users file', function (done) {
       nock.cleanAll();
+      var boundary = null;
 
       var request = nock(API_URL)
-        .post('/emails/provider')
-        .matchHeader('Authorization', 'Bearer ' + this.token)
-        .reply(200)
+        .matchHeader('Content-Type', function (header) {
+          boundary = '--' + header.match(/boundary=([^\n]*)/)[1];
 
-      this
-        .emailProvider
-        .configure(data)
-        .then(function () {
-          expect(request.isDone())
-            .to.be.true;
+          return true;
+        })
+        .post('/jobs/users-imports', function (body) {
+          var parts = extractParts(body, boundary);
 
-          done();
-        });
-    });
-  });
-
-
-  describe('#update', function () {
-    var data = {
-      name: 'Test provider',
-      credentials: {}
-    };
-
-    beforeEach(function () {
-      this.request = nock(API_URL)
-        .patch('/emails/provider')
-        .reply(200, data);
-    })
-
-
-    it('should accept a callback', function (done) {
-      this
-        .emailProvider
-        .update({}, data, function () {
-          done();
-        });
-    });
-
-
-    it('should return a promise if no callback is given', function (done) {
-      this
-        .emailProvider
-        .update({}, data)
-        .then(done.bind(null, null))
-        .catch(done.bind(null, null));
-    });
-
-
-    it('should pass any errors to the promise catch handler', function (done) {
-      nock.cleanAll();
-
-      var request = nock(API_URL)
-        .patch('/emails/provider/' + data.id)
-        .reply(500);
-
-      this
-        .emailProvider
-        .update({}, data)
-        .catch(function (err) {
-          expect(err)
+          // Validate the connection id.
+          expect(parts.connection_id)
             .to.exist
-            .to.be.an.instanceOf(Error);
+            .to.be.a('string')
+            .to.equal(data.connection_id);
+
+          // Validate the content type of the users JSON.
+          expect(parts.users)
+            .to.exist
+            .to.be.a('string')
+            .to.contain('Content-Type: application/json');
+
+          // Validate the content of the users JSON.
+          expect(parts.users.slice(-2))
+            .to.equal('[]');
+
+          return true;
+        })
+        .reply(200);
+
+      this
+        .jobs
+        .importUsers(data)
+        .then(function () {
+          expect(request.isDone())
+            .to.be.true;
 
           done();
         });
     });
 
 
-    it('should perform a PATCH request to /api/v2/emails/provider', function (done) {
+    it('should include the token in the Authorization header', function (done) {
+      nock.cleanAll();
+
+      var request = nock(API_URL)
+        .post('/jobs/users-imports')
+        .matchHeader('Authorization', 'Bearer ' + this.token)
+        .reply(200)
+
+      this
+        .jobs
+        .importUsers(data)
+        .then(function () {
+          expect(request.isDone())
+            .to.be.true;
+
+          done();
+        });
+    });
+  });
+
+
+  describe('#verifyEmail', function () {
+    var data = {
+      user_id: 'github|12345',
+    };
+
+    beforeEach(function () {
+      this.request = nock(API_URL)
+        .post('/jobs/verification-email')
+        .reply(200, data);
+    })
+
+
+    it('should accept a callback', function (done) {
+      this
+        .jobs
+        .verifyEmail(data, function () {
+          done();
+        });
+    });
+
+
+    it('should return a promise if no callback is given', function (done) {
+      this
+        .jobs
+        .verifyEmail(data)
+        .then(done.bind(null, null))
+        .catch(done.bind(null, null));
+    });
+
+
+    it('should pass any errors to the promise catch handler', function (done) {
+      nock.cleanAll();
+
+      var request = nock(API_URL)
+        .post('/jobs/verification-email')
+        .reply(500);
+
+      this
+        .jobs
+        .verifyEmail(data)
+        .catch(function (err) {
+          expect(err)
+            .to.exist;
+
+          done();
+        });
+    });
+
+
+    it('should perform a POST request to /api/v2/jobs/verification-email', function (done) {
       var request = this.request;
 
       this
-        .emailProvider
-        .update({}, data)
+        .jobs
+        .verifyEmail(data)
         .then(function () {
           expect(request.isDone())
             .to.be.true;
@@ -359,12 +401,12 @@ describe('EmailProviderManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .patch('/emails/provider', data)
+        .post('/jobs/verification-email', data)
         .reply(200);
 
       this
-        .emailProvider
-        .update({}, data)
+        .jobs
+        .verifyEmail(data)
         .then(function () {
           expect(request.isDone())
             .to.be.true;
@@ -378,91 +420,13 @@ describe('EmailProviderManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .patch('/emails/provider')
+        .post('/jobs/verification-email')
         .matchHeader('Authorization', 'Bearer ' + this.token)
         .reply(200)
 
       this
-        .emailProvider
-        .update({}, data)
-        .then(function () {
-          expect(request.isDone())
-            .to.be.true;
-
-          done();
-        });
-    });
-  });
-
-
-  describe('#delete', function () {
-    beforeEach(function () {
-      this.request = nock(API_URL)
-        .delete('/emails/provider')
-        .reply(200);
-    });
-
-
-    it('should accept a callback', function (done) {
-      this
-        .emailProvider
-        .delete({}, done.bind(null, null));
-    });
-
-
-    it('should return a promise when no callback is given', function (done) {
-      this
-        .emailProvider
-        .delete()
-        .then(done.bind(null, null));
-    });
-
-
-    it('should perform a DELETE request to /emails/provider', function (done) {
-      var request = this.request;
-
-      this
-        .emailProvider
-        .delete()
-        .then(function () {
-          expect(request.isDone())
-            .to.be.true;
-
-          done();
-        });
-    });
-
-
-    it('should pass any errors to the promise catch handler', function (done) {
-      nock.cleanAll();
-
-      var request = nock(API_URL)
-        .delete('/emails/provider')
-        .reply(500);
-
-      this
-        .emailProvider
-        .delete()
-        .catch(function (err) {
-          expect(err)
-            .to.exist;
-
-          done();
-        });
-    });
-
-
-    it('should include the token in the Authorization header', function (done) {
-      nock.cleanAll();
-
-      var request = nock(API_URL)
-        .delete('/emails/provider')
-        .matchHeader('Authorization', 'Bearer ' + this.token)
-        .reply(200)
-
-      this
-        .emailProvider
-        .delete()
+        .jobs
+        .verifyEmail(data)
         .then(function () {
           expect(request.isDone())
             .to.be.true;
