+++ test/management/users.tests.js
@@ -1,16 +1,17 @@
 var expect = require('chai').expect;
 var nock = require('nock');
 
-var RulesManager = require('../src/RulesManager');
-var ArgumentError = require('../src/exceptions').ArgumentError;
+var SRC_DIR = '../../src';
+var API_URL = 'https://tenants.auth0.com';
 
-var API_URL = 'https://tenant.auth0.com';
+var UsersManager = require(SRC_DIR + '/management/UsersManager');
+var ArgumentError = require(SRC_DIR + '/exceptions').ArgumentError;
 
 
-describe('RulesManager', function () {
+describe('UsersManager', function () {
   before(function () {
     this.token = 'TOKEN';
-    this.rules = new RulesManager({
+    this.users = new UsersManager({
       headers: { authorization: 'Bearer ' + this.token },
       baseUrl: API_URL
     });
@@ -18,11 +19,21 @@ describe('RulesManager', function () {
 
 
   describe('instance', function () {
-    var methods = ['get', 'getAll', 'create', 'update', 'delete'];
+    var methods = [
+      'get',
+      'getAll',
+      'create',
+      'update',
+      'delete',
+      'deleteAll',
+      'unlink',
+      'link',
+      'deleteMultifactorProvider'
+    ];
 
     methods.forEach(function (method) {
       it('should have a ' + method + ' method', function () {
-        expect(this.rules[method])
+        expect(this.users[method])
           .to.exist
           .to.be.an.instanceOf(Function);
       })
@@ -32,23 +43,23 @@ describe('RulesManager', function () {
 
   describe('#constructor', function () {
     it('should error when no options are provided', function () {
-      expect(RulesManager)
+      expect(UsersManager)
         .to.throw(ArgumentError, 'Must provide manager options');
     });
 
 
     it('should throw an error when no base URL is provided', function () {
-      var client = RulesManager.bind(null, {});
+      var manager = UsersManager.bind(null, {});
 
-      expect(client)
+      expect(manager)
         .to.throw(ArgumentError, 'Must provide a base URL for the API');
     });
 
 
     it('should throw an error when the base URL is invalid', function () {
-      var client = RulesManager.bind(null, { baseUrl: '' });
+      var manager = UsersManager.bind(null, { baseUrl: '' });
 
-      expect(client)
+      expect(manager)
         .to.throw(ArgumentError, 'The provided base URL is invalid');
     });
   });
@@ -57,14 +68,14 @@ describe('RulesManager', function () {
   describe('#getAll', function () {
     beforeEach(function () {
       this.request = nock(API_URL)
-        .get('/rules')
+        .get('/users')
         .reply(200);
     })
 
 
     it('should accept a callback', function (done) {
       this
-        .rules
+        .users
         .getAll(function () {
           done();
         });
@@ -73,7 +84,7 @@ describe('RulesManager', function () {
 
     it('should return a promise if no callback is given', function (done) {
       this
-        .rules
+        .users
         .getAll()
         .then(done.bind(null, null))
         .catch(done.bind(null, null));
@@ -84,14 +95,16 @@ describe('RulesManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .get('/rules')
+        .get('/users')
         .reply(500);
 
       this
-        .rules
+        .users
         .getAll()
         .catch(function (err) {
-          expect(err).to.exist;
+          expect(err)
+            .to.exist;
+
           done();
         });
     });
@@ -102,20 +115,20 @@ describe('RulesManager', function () {
 
       var data = [{ test: true }];
       var request = nock(API_URL)
-        .get('/rules')
+        .get('/users')
         .reply(200, data);
 
       this
-        .rules
+        .users
         .getAll()
-        .then(function (credentials) {
-          expect(credentials)
+        .then(function (users) {
+          expect(users)
             .to.be.an.instanceOf(Array);
 
-          expect(credentials.length)
+          expect(users.length)
             .to.equal(data.length);
 
-          expect(credentials[0].test)
+          expect(users[0].test)
             .to.equal(data[0].test);
 
           done();
@@ -123,14 +136,16 @@ describe('RulesManager', function () {
     });
 
 
-    it('should perform a GET request to /api/v2/rules', function (done) {
+    it('should perform a GET request to /api/v2/users', function (done) {
       var request = this.request;
 
       this
-        .rules
+        .users
         .getAll()
         .then(function () {
-          expect(request.isDone()).to.be.true;
+          expect(request.isDone())
+            .to.be.true;
+
           done();
         });
     });
@@ -140,15 +155,17 @@ describe('RulesManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .get('/rules')
+        .get('/users')
         .matchHeader('Authorization', 'Bearer ' + this.token)
-        .reply(200)
+        .reply(200);
 
       this
-        .rules
+        .users
         .getAll()
         .then(function () {
-          expect(request.isDone()).to.be.true;
+          expect(request.isDone())
+            .to.be.true;
+
           done();
         });
     });
@@ -162,12 +179,12 @@ describe('RulesManager', function () {
         fields: 'test'
       };
       var request = nock(API_URL)
-        .get('/rules')
+        .get('/users')
         .query(params)
-        .reply(200)
+        .reply(200);
 
       this
-        .rules
+        .users
         .getAll(params)
         .then(function () {
           expect(request.isDone())
@@ -190,7 +207,7 @@ describe('RulesManager', function () {
       };
 
       this.request = nock(API_URL)
-        .get('/rules/' + this.data.id)
+        .get('/users/' + this.data.id)
         .reply(200, this.data);
     })
 
@@ -199,25 +216,25 @@ describe('RulesManager', function () {
       var params = { id: this.data.id };
 
       this
-        .rules
+        .users
         .get(params, done.bind(null, null));
     });
 
 
     it('should return a promise if no callback is given', function (done) {
       this
-        .rules
+        .users
         .get({ id: this.data.id })
         .then(done.bind(null, null))
         .catch(done.bind(null, null));
     });
 
 
-    it('should perform a POST request to /api/v2/rules/5', function (done) {
+    it('should perform a POST request to /api/v2/users/5', function (done) {
       var request = this.request;
 
       this
-        .rules
+        .users
         .get({ id: this.data.id })
         .then(function () {
           expect(request.isDone())
@@ -232,11 +249,11 @@ describe('RulesManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .get('/rules/' + this.data.id)
+        .get('/users/' + this.data.id)
         .reply(500);
 
       this
-        .rules
+        .users
         .get({ id: this.data.id })
         .catch(function (err) {
           expect(err)
@@ -251,12 +268,12 @@ describe('RulesManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .get('/rules/' + this.data.id)
+        .get('/users/' + this.data.id)
         .matchHeader('Authorization', 'Bearer ' + this.token)
         .reply(200);
 
       this
-        .rules
+        .users
         .get({ id: this.data.id })
         .then(function () {
           expect(request.isDone())
@@ -279,14 +296,14 @@ describe('RulesManager', function () {
 
     beforeEach(function () {
       this.request = nock(API_URL)
-        .post('/rules')
+        .post('/users')
         .reply(200);
     })
 
 
     it('should accept a callback', function (done) {
       this
-        .rules
+        .users
         .create(data, function () {
           done();
         });
@@ -295,7 +312,7 @@ describe('RulesManager', function () {
 
     it('should return a promise if no callback is given', function (done) {
       this
-        .rules
+        .users
         .create(data)
         .then(done.bind(null, null))
         .catch(done.bind(null, null));
@@ -306,11 +323,11 @@ describe('RulesManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .post('/rules')
+        .post('/users')
         .reply(500);
 
       this
-        .rules
+        .users
         .create(data)
         .catch(function (err) {
           expect(err)
@@ -321,11 +338,11 @@ describe('RulesManager', function () {
     });
 
 
-    it('should perform a POST request to /api/v2/rules', function (done) {
+    it('should perform a POST request to /api/v2/users', function (done) {
       var request = this.request;
 
       this
-        .rules
+        .users
         .create(data)
         .then(function () {
           expect(request.isDone())
@@ -340,11 +357,11 @@ describe('RulesManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .post('/rules', data)
+        .post('/users', data)
         .reply(200);
 
       this
-        .rules
+        .users
         .create(data)
         .then(function () {
           expect(request.isDone())
@@ -359,12 +376,12 @@ describe('RulesManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .post('/rules')
+        .post('/users')
         .matchHeader('Authorization', 'Bearer ' + this.token)
-        .reply(200)
+        .reply(200);
 
       this
-        .rules
+        .users
         .create(data)
         .then(function () {
           expect(request.isDone())
@@ -381,32 +398,32 @@ describe('RulesManager', function () {
       this.data = { id: 5 };
 
       this.request = nock(API_URL)
-        .patch('/rules/' + this.data.id)
+        .patch('/users/' + this.data.id)
         .reply(200, this.data);
     });
 
 
     it('should accept a callback', function (done) {
       this
-        .rules
+        .users
         .update({ id: 5 }, {}, done.bind(null, null));
     });
 
 
     it('should return a promise if no callback is given', function (done) {
       this
-        .rules
+        .users
         .update({ id: 5 }, {})
         .then(done.bind(null, null))
         .catch(done.bind(null, null));
     });
 
 
-    it('should perform a PATCH request to /api/v2/rules/5', function (done) {
+    it('should perform a PATCH request to /api/v2/users/5', function (done) {
       var request = this.request;
 
       this
-        .rules
+        .users
         .update({ id: 5 }, {})
         .then(function () {
           expect(request.isDone())
@@ -421,11 +438,11 @@ describe('RulesManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .patch('/rules/' + this.data.id, this.data)
+        .patch('/users/' + this.data.id, this.data)
         .reply(200);
 
       this
-        .rules
+        .users
         .update({ id: 5 }, this.data)
         .then(function () {
           expect(request.isDone())
@@ -440,11 +457,11 @@ describe('RulesManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .patch('/rules/' + this.data.id)
+        .patch('/users/' + this.data.id)
         .reply(500);
 
       this
-        .rules
+        .users
         .update({ id: this.data.id }, this.data)
         .catch(function (err) {
           expect(err)
@@ -461,31 +478,31 @@ describe('RulesManager', function () {
 
     beforeEach(function () {
       this.request = nock(API_URL)
-        .delete('/rules/' + id)
+        .delete('/users/' + id)
         .reply(200);
     });
 
 
     it('should accept a callback', function (done) {
       this
-        .rules
+        .users
         .delete({ id: id }, done.bind(null, null));
     });
 
 
     it('should return a promise when no callback is given', function (done) {
       this
-        .rules
+        .users
         .delete({ id: id })
         .then(done.bind(null, null));
     });
 
 
-    it('should perform a delete request to /rules/' + id, function (done) {
+    it('should perform a delete request to /users/' + id, function (done) {
       var request = this.request;
 
       this
-        .rules
+        .users
         .delete({ id: id })
         .then(function () {
           expect(request.isDone())
@@ -500,11 +517,11 @@ describe('RulesManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .delete('/rules/' + id)
+        .delete('/users/' + id)
         .reply(500);
 
       this
-        .rules
+        .users
         .delete({ id: id })
         .catch(function (err) {
           expect(err)
@@ -519,13 +536,292 @@ describe('RulesManager', function () {
       nock.cleanAll();
 
       var request = nock(API_URL)
-        .delete('/rules/' + id)
+        .delete('/users/' + id)
         .matchHeader('authorization', 'Bearer ' + this.token)
-        .reply(200)
+        .reply(200);
 
       this
-        .rules
+        .users
         .delete({ id: id })
+        .then(function () {
+          expect(request.isDone())
+            .to.be.true;
+
+          done();
+        });
+    });
+  });
+
+
+  describe('#link', function () {
+    var userId = 'USER_ID';
+    var data = {
+      provider: "twitter",
+      user_id: "191919191919191",
+    };
+
+    beforeEach(function () {
+      this.request = nock(API_URL)
+        .post('/users/' + userId + '/identities')
+        .reply(200);
+    })
+
+
+    it('should accept a callback', function (done) {
+      this
+        .users
+        .link(userId, data, function () {
+          done();
+        });
+    });
+
+
+    it('should return a promise if no callback is given', function (done) {
+      this
+        .users
+        .link(userId, data)
+        .then(done.bind(null, null))
+        .catch(done.bind(null, null));
+    });
+
+
+    it('should pass any errors to the promise catch handler', function (done) {
+      nock.cleanAll();
+
+      var request = nock(API_URL)
+        .post('/users/' + userId + '/identities')
+        .reply(500);
+
+      this
+        .users
+        .link(userId, data)
+        .catch(function (err) {
+          expect(err)
+            .to.exist;
+
+          done();
+        });
+    });
+
+
+    it('should perform a POST request to /api/v2/users', function (done) {
+      var request = this.request;
+
+      this
+        .users
+        .link(userId, data)
+        .then(function () {
+          expect(request.isDone())
+            .to.be.true;
+
+          done();
+        });
+    });
+
+
+    it('should pass the data in the body of the request', function (done) {
+      nock.cleanAll();
+
+      var request = nock(API_URL)
+        .post('/users/' + userId + '/identities', data)
+        .reply(200);
+
+      this
+        .users
+        .link(userId, data)
+        .then(function () {
+          expect(request.isDone())
+            .to.be.true;
+
+          done();
+        });
+    });
+
+
+    it('should include the token in the Authorization header', function (done) {
+      nock.cleanAll();
+
+      var request = nock(API_URL)
+        .post('/users/' + userId + '/identities')
+        .matchHeader('Authorization', 'Bearer ' + this.token)
+        .reply(200);
+
+      this
+        .users
+        .link(userId, data)
+        .then(function () {
+          expect(request.isDone())
+            .to.be.true;
+
+          done();
+        });
+    });
+  });
+
+
+  describe('#unlink', function () {
+    var data = {
+      id: 'u1',
+      user_id: 'u2',
+      provider: 'auth0'
+    };
+    var url = (
+      '/users/' + data.id + '/identities/' + data.provider + '/' + data.user_id
+    );
+
+
+    beforeEach(function () {
+      this.request = nock(API_URL)
+        .delete(url)
+        .reply(200);
+    });
+
+
+    it('should accept a callback', function (done) {
+      this
+        .users
+        .unlink(data, done.bind(null, null));
+    });
+
+
+    it('should return a promise when no callback is given', function (done) {
+      this
+        .users
+        .unlink(data)
+        .then(done.bind(null, null));
+    });
+
+
+    it('should perform a DELETE request to ' + url, function (done) {
+      var request = this.request;
+
+      this
+        .users
+        .unlink(data)
+        .then(function () {
+          expect(request.isDone())
+            .to.be.true;
+
+          done();
+        });
+    });
+
+
+    it('should pass any errors to the promise catch handler', function (done) {
+      nock.cleanAll();
+
+      var request = nock(API_URL)
+        .delete(url)
+        .reply(500);
+
+      this
+        .users
+        .unlink(data)
+        .catch(function (err) {
+          expect(err)
+            .to.exist;
+
+          done();
+        });
+    });
+
+
+    it('should include the token in the authorization header', function (done) {
+      nock.cleanAll();
+
+      var request = nock(API_URL)
+        .delete(url)
+        .matchHeader('authorization', 'Bearer ' + this.token)
+        .reply(200);
+
+      this
+        .users
+        .unlink(data)
+        .then(function () {
+          expect(request.isDone())
+            .to.be.true;
+
+          done();
+        });
+    });
+  });
+
+
+  describe('#deleteMultifactorProvider', function () {
+    var data = {
+      id: 'u1',
+      provider: 'auth0'
+    };
+    var url = '/users/' + data.id + '/multifactor/' + data.provider;
+
+
+    beforeEach(function () {
+      this.request = nock(API_URL)
+        .delete(url)
+        .reply(200);
+    });
+
+
+    it('should accept a callback', function (done) {
+      this
+        .users
+        .deleteMultifactorProvider(data, done.bind(null, null));
+    });
+
+
+    it('should return a promise when no callback is given', function (done) {
+      this
+        .users
+        .deleteMultifactorProvider(data)
+        .then(done.bind(null, null));
+    });
+
+
+    it('should perform a DELETE request to ' + url, function (done) {
+      var request = this.request;
+
+      this
+        .users
+        .deleteMultifactorProvider(data)
+        .then(function () {
+          expect(request.isDone())
+            .to.be.true;
+
+          done();
+        });
+    });
+
+
+    it('should pass any errors to the promise catch handler', function (done) {
+      nock.cleanAll();
+
+      var request = nock(API_URL)
+        .delete(url)
+        .reply(500);
+
+      this
+        .users
+        .deleteMultifactorProvider(data)
+        .catch(function (err) {
+          expect(err)
+            .to.exist;
+
+          done();
+        });
+    });
+
+
+    it('should include the token in the authorization header', function (done) {
+      nock.cleanAll();
+
+      var request = nock(API_URL)
+        .delete(url)
+        .matchHeader('authorization', 'Bearer ' + this.token)
+        .reply(200);
+
+      this
+        .users
+        .deleteMultifactorProvider(data)
         .then(function () {
           expect(request.isDone())
             .to.be.true;
